+++ src/org/netbeans/modules/php/blade/syntax/antlr4/v10/ColoringLexerAdaptor.java
@@ -128,3 +128,4 @@ public abstract class ColoringLexerAdaptor extends Lexer {
         this.more();
     }
 }
+